Added ability for wakes to have different lengths
The wake constructor now has an overload that takes in an array of wake
lengths. This allows for upstream rotors to trail more wake than
downstream rotor. This also nececitated changes in the vtu export
functions.

There was also some work done to perform some memory cleanup of the
vtu stuff so that we don't constantly leak out vtk objects.