Added a simple wing model
The simple wing model simulates a wing as a single
vortex filement with it's strength computed by a provided
lift coefficient. It uses the Kutta-Jukowski theorem to
then compute the vortex circulation strength.

The SimpleWing class adheres to the Inflow interface and can
be added as an additional inflow model so the wing can effect
wake advection.

Any SimpleWing inflow model must be added at the end of the inflow
array, after any rotor inflow models, for the code to work properly.
Super hacky.