New Change of Basis/Frame system. Fixed wake normalization.
A brand new frame of reference/change of basis system is now
implemented. This allows for complex vehicle geometries to
be simulated. Changes have been mode to both OpenCOPTER and
OC Fly to support this. The OC Fly updates also convert the
OpenCOPTER Frames to WOPWOP change of basis blocks.

The circulation strengths that make up the wake are no longer
normalized. This is to support multiple rotors that may have
different sizes and speeds.