Updates to openvsp support and some bug fixes
We now support OpenVSP hinges. Most of this comes in the form
of a new motion type called 'static'. This allows us to set
the angle of the hinge in the parameter file, instead of
needing multiple openvsp files for every different tilt
angle.

Fixed an issue in the wake advection, where we were
treating a vector as if it were a point when doing
coordinate tranforms. This was erroneously adding
extra advection velocity components when rotating
back into global space.

Fixed an issue where we weren't replicating OpenVSP
rotations correctly.