Fixes and changes courtesy of BRRC
Fixed long standing issue with observers following
the vehicle being backwards.

Fixed another issue where we were basing the iter_per_rev
and the observer time range on the fastest rotor in the
system. This is good for setting the timestep, but not the
total run time. Now we base iter_per_rev and all other
per rev things off the slowest rotor so that we ensure
we always capture the minimum amount of data for the
slowest rotor in the system.