Some fixes for vsp3 files, oc_fly fixes
The fix for airfoil extents wasn't implemented for
vsp3 files, it now is. This should maybe be lifted
to it's own function.

An erroneous indent was preventing some data from
being saved in the results.mat file as it was
being blocked by an incorrect flag.

Some fixes to the vsp3 input files. Added new loading
noise acoustics file.

Fixed the sphere_observer file.